backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
pyjwt
pytest
pytest-asyncio
pytest-mock
httpx
ruff
gurobipy
pandas
requests